Time: 1 ms (100.00%), Space: 51.7 MB (82.30%) - LeetHub

+++ 53-maximum-subarray/53-maximum-subarray.java
@@ -1,6 +1,6 @@
 class Solution {
     public int maxSubArray(int[] nums) {
-     long max=nums[0],s=nums[0];
+     int max=nums[0],s=nums[0];
      for(int i=1;i<nums.length;i++)
      {
           if(s<0)
@@ -9,6 +9,6 @@ class Solution {
          if(s>max)
              max=s;
      }
-        return (int)max;
+    return max;
     }
 }
